First working version.
Enough things work for this code to be functional.  Basically, buttons can be
created and clicked on.  Rollovers also work, which is a nice feature.

However, there are still some significant problems.  The button bar itself is
invisible and does not place the buttons very well.  Furthermore, text is
poorly placed within the buttons themselves.